Defer guessing the key until it is actually needed
This allows the user to explicitly set the song key and prevents
the transposer to always try to guess the song key during construction.

A use case where this would fail would be the sequence `G#` in the key
of `C#m`, because `G#` is considered an invalid chord on its own.

Fixes #6