Add more reliable guessing for key signatures.
When the first chord is not a valid key signature, the signature is
guessed by finding any key with that chord in it.

This change also fixes a bug where the wrong enharmonic chord was used
in some key signatures. For example, F# major would use F when the
correct chord is E#. Each key signature now stores a chromatic scale
with the correct pitches.

Also switched to jasmine for unit tests.

Fixes #13